tests/unit/test_model.py
"""
Unit tests for model training logic.
"""
import pytest
from unittest.mock import Mock, MagicMock
import pandas as pd
import numpy as np


def test_churn_model_initialization():
    """Test ChurnModel initialization."""
    from src.models.churn_model import ChurnModel
    
    model = ChurnModel(catalog="test_catalog", schema="test_schema")
    
    assert model.catalog == "test_catalog"
    assert model.schema == "test_schema"
    assert model.model is None


def test_create_training_set_structure():
    """Test training set creation returns expected structure."""
    from src.models.churn_model import ChurnModel
    
    # This would require mocking Feature Store client
    # For demonstration, showing test structure
    pass


def test_model_training_metrics():
    """Test model training produces expected metrics."""
    from sklearn.linear_model import LogisticRegression
    
    # Create synthetic data
    np.random.seed(42)
    X = np.random.rand(100, 5)
    y = np.random.randint(0, 2, 100)
    
    model = LogisticRegression(random_state=42)
    model.fit(X, y)
    
    # Verify model is trained
    assert hasattr(model, 'coef_')
    assert model.coef_.shape[1] == 5